+++ ToDoList/Task.swift
@@ -11,8 +11,24 @@ import RealmSwift
 class Task: Object, ObjectKeyIdentifiable {
     @Persisted(primaryKey: true) var id: ObjectId
     @Persisted var title: String
-    @Persisted var details: String = ""
+    @Persisted var detail: String = ""
     @Persisted var date: Date = Date()
-    @Persisted var descriptionVisibility: Bool = true
-    @Persisted var isComplete: Bool = false
+    @Persisted var detailVisibility: Bool = true
+    @Persisted var isCompleted: Bool = false
+    
+//    init(_ title: String) {
+//        self.title = title
+////        detail = ""
+////        date = Date()
+////        detailVisibility = true
+////        isCompleted = false
+//    }
+//
+//    init(_ title: String, _ detail: String, _ date: Date, _ detailVisibility: Bool, _ isCompleted: Bool) {
+//        self.title = title
+//        self.detail = detail
+//        self.date = date
+//        self.detailVisibility = detailVisibility
+//        self.isCompleted = isCompleted
+//    }
 }
