+++ ToDoList/Views/Home.swift
@@ -8,8 +8,7 @@
 import SwiftUI
 
 struct Home: View {
-    
-    @EnvironmentObject var taskViewModel: TaskViewModel
+    @StateObject var realmManager = RealmManager()
     @State private var showModal = false
     @Namespace var animation
     @State var currentDate: Date = Date()
@@ -18,81 +17,85 @@ struct Home: View {
         NavigationView {
             VStack {
                 HeaderView()
-                    if let filtered = taskViewModel.filteredTasks {
-                        if filtered.isEmpty {
-                            NoTaskView()
-                        }
-                        else {
-                            List {
-                                Section {
-                                    ForEach(filtered) { task in
-                                        if !task.isComplete {
-                                            TaskCardView(task: task)
-                                                .swipeActions(edge: .leading) {
-                                                    Button (action: { taskViewModel.updateTaskCompletion(task: task) }) {
-                                                        Label("Done", systemImage: "checkmark")
-                                                    }
-                                                    .tint(.green)
-                                                }
-                                            
-                                                .onTapGesture {
-                                                    self.showModal = true
-                                                }
-                                                .swipeActions {
-                                                    Button(role: .destructive) {
-                                                        withAnimation(.linear(duration: 0.4)) {
-                                                            //taskViewModel.deleteTask(indexSet: )
-                                                            print("delete")
-                                                        }
-                                                    } label: {
-                                                        Label("Delete", systemImage: "trash")
-                                                    }
-                                                }
-                                        }
-                                    }
-
-                                    .sheet(isPresented: self.$showModal) {
-                                        ModalView()
-                                    }
-                                }
-                                
-                                Section {
-                                    ForEach(filtered) { task in
-                                        if task.isComplete {
-                                            TaskDoneCardView(task: task)
-                                                .onChange(of: taskViewModel.currentDay) { newValue in
-                                                    taskViewModel.filterTodayTasks()
-                                                }
-                                                .onTapGesture {
-                                                    self.showModal = true
-                                                }
-                                                .swipeActions {
-                                                    Button(role: .destructive) {
-                                                        withAnimation(.linear(duration: 0.4)) {
-                                                            //taskViewModel.deleteTask(indexSet: )
-                                                            print("delete")
-                                                        }
-                                                    } label: {
-                                                        Label("Delete", systemImage: "trash")
-                                                    }
-                                                }
-                                        }
-                                    }
-                                    .sheet(isPresented: self.$showModal) {
-                                        ModalView()
-                                    }
-                                    
-                                }
-                            }
-                            .frame(minHeight: 500)
-                            .background(Color.yellow)
-                        }
-                    }
+//                if let filtered = taskViewModel.filteredTasks {
+//                    if filtered.isEmpty {
+//                        NoTaskView()
+//                    }
+//                    else {
+//                        List {
+//                            Section {
+//                                ForEach(filtered) { task in
+//                                    if !task.isComplete {
+//                                        TaskCardView(task: task)
+//                                            .swipeActions(edge: .leading) {
+//                                                Button (action: { taskViewModel.updateTaskCompletion(task: task) }) {
+//                                                    Label("Done", systemImage: "checkmark")
+//                                                }
+//                                                .tint(.green)
+//                                            }
+//
+//                                            .onTapGesture {
+//                                                self.showModal = true
+//                                            }
+//                                            .swipeActions {
+//                                                Button(role: .destructive) {
+//                                                    withAnimation(.linear(duration: 0.4)) {
+//                                                        //taskViewModel.deleteTask(indexSet: )
+//                                                        print("delete")
+//                                                    }
+//                                                } label: {
+//                                                    Label("Delete", systemImage: "trash")
+//                                                }
+//                                            }
+//                                    }
+//                                }
+//
+//                                .sheet(isPresented: self.$showModal) {
+//                                    ModalView()
+//                                        .environmentObject(realmManager)
+//                                }
+//                            }
+//
+//                            Section {
+//                                ForEach(filtered) { task in
+//                                    if task.isComplete {
+//                                        TaskDoneCardView(task: task)
+//                                            .onChange(of: taskViewModel.currentDay) { newValue in
+//                                                taskViewModel.filterTodayTasks()
+//                                            }
+//                                            .onTapGesture {
+//                                                self.showModal = true
+//                                            }
+//                                            .swipeActions {
+//                                                Button(role: .destructive) {
+//                                                    withAnimation(.linear(duration: 0.4)) {
+//                                                        //taskViewModel.deleteTask(indexSet: )
+//                                                        print("delete")
+//                                                    }
+//                                                } label: {
+//                                                    Label("Delete", systemImage: "trash")
+//                                                }
+//                                            }
+//                                    }
+//                                }
+//                                .sheet(isPresented: self.$showModal) {
+//                                    ModalView()
+//                                        .environmentObject(realmManager)
+//                                }
+//
+//                            }
+//                        }
+//                        .frame(minHeight: 500)
+//                        .background(Color.yellow)
+//                    }
+//                }
+//                Spacer()
             }
             .navigationTitle("TEAM SUNA")
-            .navigationBarItems(trailing: NavigationLink("Calendar", destination: CalendarView(currentDate: $currentDate)))
+            .navigationBarItems(trailing: NavigationLink("Calendar", destination: CalendarView(currentDate: $currentDate)
+                .environmentObject(realmManager)))
         }
-        .environmentObject(TaskViewModel())
+        .environmentObject(realmManager)
         .safeAreaInset(edge: .bottom) {
             Button {
                 self.showModal = true
@@ -104,6 +107,7 @@ struct Home: View {
             }
             .sheet(isPresented: self.$showModal) {
                 ModalView()
+                    .environmentObject(realmManager)
             }
             .padding(.horizontal)
             .padding(.top, 10)
@@ -115,13 +119,8 @@ struct Home: View {
 }
 
 struct Home_Previews: PreviewProvider {
-    
-    //    @EnvironmentObject var taskViewModel: TaskViewModel
-    
     static var previews: some View {
         Home()
-            .environmentObject(TaskViewModel())
-        
     }
 }
