+++ ToDoList/Views/ModalView.swift
@@ -10,18 +10,17 @@ import SwiftUI
 struct ModalView: View {
     @Environment(\.presentationMode) var presentation
     @Environment(\.dismiss) private var dismiss
-    @EnvironmentObject var taskViewModel: TaskViewModel
-
+    @EnvironmentObject var realmManager: RealmManager
     
     //if let으로 초기화를 해서 모델이 하나 들어올 시 해당 모델의 값으로 변수들을 초기화, 그렇지 않다면 그냥 기본값으로 초기화
     
     
     // MARK: Task values
-    @State var taskTitle: String = ""
-    @State var taskDescription: String = ""
-    @State var taskDate: Date = Date()
-    @State var taskDone: Bool = false
-    @State var descriptionVisible: Bool = true
+    @State var title: String = ""
+    @State var detail: String = ""
+    @State var date: Date = Date()
+    @State var isCompleted: Bool = false
+    @State var detailVisible: Bool = true
     
     var body: some View {
         
@@ -29,12 +28,12 @@ struct ModalView: View {
             VStack {
                 List {
                     Section {
-                        TextField("task", text: $taskTitle)
-                        TextField("description", text: $taskDescription)
+                        TextField("task", text: $title)
+                        TextField("description", text: $detail)
                         HStack {
                             Text("time")
                             Spacer()
-                            DatePicker("", selection: $taskDate)
+                            DatePicker("", selection: $date)
                                 .datePickerStyle(.compact)
                                 .labelsHidden()
                         }
@@ -42,22 +41,21 @@ struct ModalView: View {
                         Text("Task")
                     }
                     Section {
-                        Toggle("Done", isOn: $taskDone)
-                        Toggle("display description", isOn: $descriptionVisible)
+                        Toggle("Done", isOn: $isCompleted)
+                        Toggle("display description", isOn: $detailVisible)
                     } header: {
                         Text("option")
                     }
                 }
                 .listStyle(.grouped)
                 .navigationBarTitleDisplayMode(.inline)
-                .interactiveDismissDisabled()
                 .toolbar {
                     ToolbarItem(placement: .navigationBarTrailing) {
                         Button("save") {
-                            taskViewModel.addTask(title: taskTitle, description: taskDescription, date: taskDate, descriptionVisibility: descriptionVisible, isComplete: taskDone)
+                            realmManager.addTask(title, detail, date, detailVisible, isCompleted)
                             dismiss()
                         }
-                        .disabled(taskTitle == "")
+                        .disabled(title == "")
                     }
                     ToolbarItem(placement: .navigationBarLeading) {
                         Button("cancel") {
